Local Devices & Path Fix
This patch fixes a bug where local devices were no longer being sent
with the API calls that required them. Additionally, the logic to arrive
at the libStorage home directory has been updated so that if
LIBSTORAGE_HOME is not set, it is set to /. And if the process is
running as non-root, the home dir is changed once again to .libstorage
under the $HOME directory of the owning process user.