Support for GoIsilon v1.1.0
This patch adds support for the GoIsilon v1.1.0 release, enabling
support for custom volume paths and enhanced NFS export security
settings.

Note: This update does break backwards compatibility for the Isilon
driver's `volumePath` configuration key. This key's value used to be
appended to the base path `/ifs/volumes`, but that's no longer the case.
The value of the `volumePath` key is now considered to be absolute. If
it is not provided or empty, the value `/ifs/volumes` will be used. But
if a value is specified for the key, that is the value used for the
volume path, sans any prefix.